+++ tests/TestCore.py
@@ -75,6 +75,14 @@ class TestCore:
         if not want_argv:
             del out['vars']['argv']
         del out['vars']['pashmakinfo']
+        try:
+            del out['vars']['__file__']
+        except:
+            pass
+        try:
+            del out['vars']['__dir__']
+        except:
+            pass
 
         return out