+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### Added
 - added module path system
+- added `$__file__` and `$__dir__` general variables
 
 #### Fixed
 - raise error when namespace or function name contains `.` character